HuggingFace_RAG_LangChainGraph/requirements.txt
streamlit
langchain
langgraph
openai
faiss-cpu
pydantic
tiktoken
langchain_community
python-dotenv
langchain-openai
langchain-community
grandalf
